ingestion-service/requirements.txt
boto3==1.34.162
psycopg2-binary==2.9.9
python-dotenv==1.0.1
